Added RMSNorm to LayerNormMLP
